+++ CHANGES.md
@@ -1,4 +1,7 @@
 
+
+
+- Add `Fpath.mem_ext`.
 - Documentation fixes.
 
 v0.7.0 2016-05-23 La Forclaz (VS)